polousers/requirements.txt
cffi==0.9.2
characteristic==14.3.0
cryptography==0.8.2
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.12
pyOpenSSL==0.15.1
service-identity==14.0.0
six==1.9.0
Twisted==15.1.0
zope.interface==4.1.2
